requirements.txt
loguru==0.7.3
python-dotenv==1.0.1
PyYAML==6.0.2
roboflow==1.1.53
setuptools==57.4.0
setuptools==75.8.0
torch==2.5.1+cu118
tqdm==4.67.1
typer==0.15.1
ultralytics==8.3.68
wandb==0.19.6